MedbucketAPI/requirements.txt
asgiref==3.2.10
boto3==1.16.50
botocore==1.19.50
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-storages==1.11.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.4
reportlab==3.5.59
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
mysqlclient==2.0.3